Refactored to replace the resource soft reserved word
