+++ tests/BaseTest.php
@@ -19,7 +19,7 @@ abstract class BaseTest extends \PHPUnit_Framework_TestCase
     {
         $this->paylike        = new Paylike("dbcf01af-8667-4967-9791-56101ca87ac8");
         $this->app_id         = "594d3cde5be12d547cbe2ec2";
-        $this->transaction_id = "594d3d2cbe4728547d40150e";
+        $this->transaction_id = "5da8272132aad22568a511b7";
         $this->merchant_id    = "594d3c455be12d547cbe2ebe";
     }
 }
